fix by reducing time to refocus
